Making Cache size configurable

+++ blobfuse/include/DataLakeBfsClient.h
@@ -73,7 +73,7 @@ private:
     /// Helper function - Renames cached files
     ///</summary>
     ///<returns>Error value</return>
-    int rename_cached_file(std::string src, std::string dest);
+    long int rename_cached_file(std::string src, std::string dest);
     ///<summary>
     /// ADLS Client to make dfs storage calls
     ///</summary>
